Fix failure in android.provider.cts.MediaStore_FilesTest
openFileDescriptor or openOutputStream will call openFile in MediaProvider.
And the parent dir of the caller's url is created by openFile.
But MediaStore_FilesTest expect that the dir does not exists.
So the test case fail.

Remove the function of creating the dir to make MediaStore_FilesTest
successfully in openFile. And create the dir in the request function.

Change-Id: Ia4602e1c34ff0000ed1fa849e4e1d9c0b8a62d68